HIP13 | Morgan Gwan | extract OpenMrsTest the test into DiscoveryDataSourceTest
