+++ src/In.ProjectEKA.HipService/OpenMrs/IOpenMrsClient.cs
@@ -1,0 +1,10 @@
+using System.Net.Http;
+using System.Threading.Tasks;
+
+namespace In.ProjectEKA.HipServiceTest.OpenMrs
+{
+    public interface IOpenMrsClient
+    {
+        Task<HttpResponseMessage> GetAsync(string openmrsUrl);
+    }
+}
