+++ src/In.ProjectEKA.HipService/OpenMrs/OpenMrsClient.cs
@@ -8,7 +8,7 @@ using In.ProjectEKA.HipLibrary.Patient.Model;
 
 namespace In.ProjectEKA.HipServiceTest.OpenMrs
 {
-    public class OpenMrsClient: IPatientDal
+    public class OpenMrsClient: IOpenMrsClient
     {
         private readonly HttpClient httpClient;
         private readonly OpenMrsConfiguration configuration;
@@ -26,11 +26,6 @@ namespace In.ProjectEKA.HipServiceTest.OpenMrs
                     Convert.ToBase64String(authToken));
 
             return await httpClient.GetAsync(configuration.Url + openmrsUrl);;
-        }
-
-        public List<Hl7.Fhir.Model.Patient> LoadPatients(string name, Gender? gender, string yearOfBirth)
-        {
-            throw new NotImplementedException();
         }
     }
 }
